move selenopsCLI-s commented code to new selenopsCore Crawler class
